feat(misc): `size_t` and `I32Rect` return types replaced with `u32` and `U32Rect` accordingly

+++ engine/misc/Window.cpp
@@ -43,7 +43,7 @@ namespace Kiwi {
         return GetWindowSizes().Width();
     }
 
-    size_t AWindow::GetFramebufferWidth() const {
+    u32 AWindow::GetFramebufferWidth() const {
         return GetFramebufferSizes().Width();
     }
 
@@ -51,7 +51,7 @@ namespace Kiwi {
         return GetWindowSizes().Height();
     }
 
-    size_t AWindow::GetFramebufferHeight() const {
+    u32 AWindow::GetFramebufferHeight() const {
         return GetFramebufferSizes().Height();
     }
 }
